+++ sdk/appconfiguration/azure-appconfiguration/tests/async_wrapper.py
@@ -6,14 +6,9 @@
 # --------------------------------------------------------------------------
 from devtools_testutils import PowerShellPreparer
 
-from azure.appconfiguration import (
-    ResourceReadOnlyError,
-    ConfigurationSetting,
-)
+from azure.appconfiguration import ConfigurationSetting
 from azure.appconfiguration.aio import AzureAppConfigurationClient
-from azure.core.exceptions import (
-    AzureError,
-)
+from azure.core.exceptions import AzureError
 from consts import (
     KEY,
     LABEL,
@@ -66,7 +61,6 @@ def trim_kwargs_from_test_function(fn, kwargs):
             args = set(args)
             for key in [k for k in kwargs if k not in args]:
                 del kwargs[key]
-
 
 def app_config_decorator(func, **kwargs):