+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_business_card.py
@@ -32,14 +32,6 @@ class TestBusinessCard(FormRecognizerTest):
             poller = client.begin_recognize_business_cards(myfile)
 
     @FormRecognizerPreparer()
-    def test_authentication_successful_key(self, formrecognizer_test_endpoint, formrecognizer_test_api_key):
-        client = FormRecognizerClient(formrecognizer_test_endpoint, AzureKeyCredential(formrecognizer_test_api_key))
-        with open(self.business_card_jpg, "rb") as fd:
-            myfile = fd.read()
-        poller = client.begin_recognize_business_cards(myfile)
-        result = poller.result()
-
-    @FormRecognizerPreparer()
     def test_authentication_bad_key(self, formrecognizer_test_endpoint, formrecognizer_test_api_key):
         client = FormRecognizerClient(formrecognizer_test_endpoint, AzureKeyCredential("xxxx"))
         with self.assertRaises(ClientAuthenticationError):
@@ -251,52 +243,6 @@ class TestBusinessCard(FormRecognizerTest):
 
     @FormRecognizerPreparer()
     @GlobalClientPreparer()
-    def test_business_card_jpg(self, client):
-
-        with open(self.business_card_jpg, "rb") as fd:
-            business_card = fd.read()
-
-        poller = client.begin_recognize_business_cards(business_card)
-
-        result = poller.result()
-        self.assertEqual(len(result), 1)
-        business_card = result[0]
-        # check dict values
-        self.assertEqual(len(business_card.fields.get("ContactNames").value), 1)
-        self.assertEqual(business_card.fields.get("ContactNames").value[0].value_data.page_number, 1)
-        self.assertEqual(business_card.fields.get("ContactNames").value[0].value['FirstName'].value, 'Avery')
-        self.assertEqual(business_card.fields.get("ContactNames").value[0].value['LastName'].value, 'Smith')
-
-        self.assertEqual(len(business_card.fields.get("JobTitles").value), 1)
-        self.assertEqual(business_card.fields.get("JobTitles").value[0].value, "Senior Researcher")
-
-        self.assertEqual(len(business_card.fields.get("Departments").value), 1)
-        self.assertEqual(business_card.fields.get("Departments").value[0].value, "Cloud & Al Department")
-
-        self.assertEqual(len(business_card.fields.get("Emails").value), 1)
-        self.assertEqual(business_card.fields.get("Emails").value[0].value, "avery.smith@contoso.com")
-
-        self.assertEqual(len(business_card.fields.get("Websites").value), 1)
-        self.assertEqual(business_card.fields.get("Websites").value[0].value, "https://www.contoso.com/")
-
-        # FIXME: uncomment https://github.com/Azure/azure-sdk-for-python/issues/14300
-        # self.assertEqual(len(business_card.fields.get("MobilePhones").value), 1)
-        # self.assertEqual(business_card.fields.get("MobilePhones").value[0].value, "https://www.contoso.com/")
-
-        # self.assertEqual(len(business_card.fields.get("OtherPhones").value), 1)
-        # self.assertEqual(business_card.fields.get("OtherPhones").value[0].value, "https://www.contoso.com/")
-
-        # self.assertEqual(len(business_card.fields.get("Faxes").value), 1)
-        # self.assertEqual(business_card.fields.get("Faxes").value[0].value, "https://www.contoso.com/")
-
-        self.assertEqual(len(business_card.fields.get("Addresses").value), 1)
-        self.assertEqual(business_card.fields.get("Addresses").value[0].value, "2 Kingdom Street Paddington, London, W2 6BD")
-
-        self.assertEqual(len(business_card.fields.get("CompanyNames").value), 1)
-        self.assertEqual(business_card.fields.get("CompanyNames").value[0].value, "Contoso")
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
     def test_business_card_png(self, client):
 
         with open(self.business_card_png, "rb") as stream:
@@ -418,6 +364,40 @@ class TestBusinessCard(FormRecognizerTest):
         for name, field in business_card.fields.items():
             for f in field.value:
                 self.assertFieldElementsHasValues(f.value_data.field_elements, business_card.page_range.first_page_number)
+        
+        # check dict values
+        self.assertEqual(len(business_card.fields.get("ContactNames").value), 1)
+        self.assertEqual(business_card.fields.get("ContactNames").value[0].value_data.page_number, 1)
+        self.assertEqual(business_card.fields.get("ContactNames").value[0].value['FirstName'].value, 'Avery')
+        self.assertEqual(business_card.fields.get("ContactNames").value[0].value['LastName'].value, 'Smith')
+
+        self.assertEqual(len(business_card.fields.get("JobTitles").value), 1)
+        self.assertEqual(business_card.fields.get("JobTitles").value[0].value, "Senior Researcher")
+
+        self.assertEqual(len(business_card.fields.get("Departments").value), 1)
+        self.assertEqual(business_card.fields.get("Departments").value[0].value, "Cloud & Al Department")
+
+        self.assertEqual(len(business_card.fields.get("Emails").value), 1)
+        self.assertEqual(business_card.fields.get("Emails").value[0].value, "avery.smith@contoso.com")
+
+        self.assertEqual(len(business_card.fields.get("Websites").value), 1)
+        self.assertEqual(business_card.fields.get("Websites").value[0].value, "https://www.contoso.com/")
+
+        # FIXME: uncomment https://github.com/Azure/azure-sdk-for-python/issues/14300
+        # self.assertEqual(len(business_card.fields.get("MobilePhones").value), 1)
+        # self.assertEqual(business_card.fields.get("MobilePhones").value[0].value, "https://www.contoso.com/")
+
+        # self.assertEqual(len(business_card.fields.get("OtherPhones").value), 1)
+        # self.assertEqual(business_card.fields.get("OtherPhones").value[0].value, "https://www.contoso.com/")
+
+        # self.assertEqual(len(business_card.fields.get("Faxes").value), 1)
+        # self.assertEqual(business_card.fields.get("Faxes").value[0].value, "https://www.contoso.com/")
+
+        self.assertEqual(len(business_card.fields.get("Addresses").value), 1)
+        self.assertEqual(business_card.fields.get("Addresses").value[0].value, "2 Kingdom Street Paddington, London, W2 6BD")
+
+        self.assertEqual(len(business_card.fields.get("CompanyNames").value), 1)
+        self.assertEqual(business_card.fields.get("CompanyNames").value[0].value, "Contoso")
 
     @FormRecognizerPreparer()
     @GlobalClientPreparer()