+++ sdk/monitor/azure-monitor-opentelemetry-exporter/samples/traces/sample_trace.py
@@ -12,9 +12,14 @@ from opentelemetry.sdk.trace.export import BatchExportSpanProcessor
 from azure.monitor.opentelemetry.exporter import AzureMonitorTraceExporter
 
 
-exporter = AzureMonitorTraceExporter(
-    connection_string = os.environ["APPLICATIONINSIGHTS_CONNECTION_STRING"]
+exporter = AzureMonitorTraceExporter.from_connection_string(
+    os.environ["APPLICATIONINSIGHTS_CONNECTION_STRING"]
 )
+
+# You can also instantiate the exporter via the constructor
+# The connection string will be automatically populated via the
+# APPLICATIONINSIGHTS_CONNECTION_STRING environment variable
+# exporter = AzureMonitorTraceExporter()
 
 trace.set_tracer_provider(TracerProvider())
 tracer = trace.get_tracer(__name__)