+++ sdk/resources/azure-mgmt-resource/azure/mgmt/resource/managedapplications/aio/operations/__init__.py
@@ -6,10 +6,12 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
+from ._application_client_operations import ApplicationClientOperationsMixin
 from ._applications_operations import ApplicationsOperations
 from ._application_definitions_operations import ApplicationDefinitionsOperations
 
 __all__ = [
+    'ApplicationClientOperationsMixin',
     'ApplicationsOperations',
     'ApplicationDefinitionsOperations',
 ]