+++ sdk/textanalytics/azure-ai-textanalytics/samples/sample_analyze_healthcare_entities.py
@@ -11,8 +11,14 @@ FILE: sample_analyze_healthcare_entities.py
 
 DESCRIPTION:
     This sample demonstrates how to detect healthcare entities in a batch of documents.
-    Each entity found in the document will have a link associated with it from a
-    data source.  Relations between entities will also be included in the response.
+
+    In this sample we will be a newly-hired engineer working in a pharmacy. We are going to
+    comb through all of the prescriptions our pharmacy has fulfilled so we can catalog how
+    much inventory we have.
+
+    As a usage note: healthcare is currently in gated preview. Your subscription needs to
+    be allow-listed before you can use this endpoint. More information about that here:
+    https://aka.ms/text-analytics-health-request-access
 
 USAGE:
     python sample_analyze_healthcare_entities.py
@@ -23,15 +29,23 @@ USAGE:
 """
 
 
-import os
-
 
 class AnalyzeHealthcareEntitiesSample(object):
 
     def analyze_healthcare_entities(self):
+
+        print(
+            "In this sample we will be combing through the prescriptions our pharmacy has fulfilled "
+            "so we can catalog how much inventory we have"
+        )
+        print(
+            "We start out with a list of prescription documents."
+        )
+
         # [START analyze_healthcare_entities]
+        import os
         from azure.core.credentials import AzureKeyCredential
-        from azure.ai.textanalytics import TextAnalyticsClient
+        from azure.ai.textanalytics import TextAnalyticsClient, HealthcareEntityRelationType
 
         endpoint = os.environ["AZURE_TEXT_ANALYTICS_ENDPOINT"]
         key = os.environ["AZURE_TEXT_ANALYTICS_KEY"]
@@ -42,15 +56,21 @@ class AnalyzeHealthcareEntitiesSample(object):
         )
 
         documents = [
-            "Subject is taking 100mg of ibuprofen twice daily"
+            """
+            Patient needs to take 100 mg of ibuprofen, and 3 mg of potassium. Also needs to take
+            10 mg of Zocor.
+            """,
+            """
+            Patient needs to take 50 mg of ibuprofen, and 2 mg of Coumadin.
+            """
         ]
 
-        poller = text_analytics_client.begin_analyze_healthcare_entities(documents, show_stats=True)
+        poller = text_analytics_client.begin_analyze_healthcare_entities(documents)
         result = poller.result()
 
         docs = [doc for doc in result if not doc.is_error]
 
-        print("Results of Healthcare Entities Analysis:")
+        print("Let's first visualize the outputted healthcare result:")
         for idx, doc in enumerate(docs):
             for entity in doc.entities:
                 print("Entity: {}".format(entity.text))
@@ -63,14 +83,50 @@ class AnalyzeHealthcareEntitiesSample(object):
                     for data_source in entity.data_sources:
                         print("......Entity ID: {}".format(data_source.entity_id))
                         print("......Name: {}".format(data_source.name))
-                if len(entity.related_entities) > 0:
-                    print("...Related Entities:")
-                    for related_entity, relation_type in entity.related_entities.items():
-                        print("......Entity Text: {}".format(related_entity.text))
-                        print("......Relation Type: {}".format(relation_type))
+            for relation in doc.entity_relations:
+                print("Relation of type: {} has the following roles".format(relation.relation_type))
+                for role in relation.roles:
+                    print("...Role '{}' with entity '{}'".format(role.name, role.entity.text))
             print("------------------------------------------")
 
+        print("Now, let's get all of medication dosage relations from the documents")
+        dosage_of_medication_relations = [
+            entity_relation
+            for doc in docs
+            for entity_relation in doc.entity_relations if entity_relation.relation_type == HealthcareEntityRelationType.DOSAGE_OF_MEDICATION
+        ]
         # [END analyze_healthcare_entities]
+
+        print(
+            "Now, I will create a dictionary of medication to total dosage. "
+            "I will use a regex to extract the dosage amount. For simplicity sake, I will assume "
+            "all dosages are represented with numbers and have mg unit."
+        )
+        import re
+        from collections import defaultdict
+
+        medication_to_dosage = defaultdict(int)
+
+        for relation in dosage_of_medication_relations:
+            # The DosageOfMedication relation should only contain the dosage and medication roles
+
+            dosage_role = next(filter(lambda x: x.name == "Attribute", relation.roles))
+            medication_role = next(filter(lambda x: x.name == "Entity", relation.roles))
+
+            try:
+                dosage_value = int(re.findall(r"\d+", dosage_role.entity.text)[0]) # we find the numbers in the dosage
+                medication_to_dosage[medication_role.entity.text] += dosage_value
+            except StopIteration:
+                # Error handling for if there's no dosage in numbers.
+                pass
+
+        [
+            print("We have fulfilled '{}' total mg of '{}'".format(
+                dosage, medication
+            ))
+            for medication, dosage in medication_to_dosage.items()
+        ]
+
 
 if __name__ == "__main__":
     sample = AnalyzeHealthcareEntitiesSample()