+++ tools/azure-devtools/src/azure_devtools/perfstress_tests/random_stream.py
@@ -43,12 +43,17 @@ class RandomStream:
         self._remaining = self._remaining - e
         self._position += e
         return self._base_data[:e]
-    
+
     def tell(self):
         return self._position
 
-    def seek(self, index):
-        self._position = index
+    def seek(self, index, whence=0):
+        if whence == 0:
+            self._position = index
+        elif whence == 1:
+            self._position = self._position + index
+        elif whence == 2:
+            self._position = self._length - 1 + index
 
     def remaining(self):
         return self._remaining
@@ -72,6 +77,6 @@ class WriteStream:
 
     def seekable(self):
         return True
-    
+
     def tell(self):
         return self._position